+++ packages/terra-legacy-theme/tests/wdio/theme-spec.js
@@ -34,17 +34,4 @@ describe('Legacy Theme', () => {
     Terra.should.beAccessible({ context: '[data-routing-menu]' });
     Terra.should.matchScreenshot({ selector: '[data-routing-menu]' });
   });
-
-  describe('Navigation Side Menu Selected Hover', () => {
-    beforeEach(() => {
-      browser.url('/#/site/components');
-      browser.click('[data-menu-item="Site"]');
-      browser.click('button[aria-label="Back"]');
-      browser.execute('document.body.classList.add("terra-legacy-theme");');
-      browser.moveToObject('[data-menu-item="Site"]');
-    });
-
-    Terra.should.beAccessible({ context: '[data-routing-menu]' });
-    Terra.should.matchScreenshot({ selector: '[data-routing-menu]' });
-  });
 });