+++ packages/terra-form-textarea/CHANGELOG.md
@@ -4,6 +4,11 @@ ChangeLog
 Unreleased
 ----------
 
+2.11.0 - (May 25, 2018)
+------------------
+### Changed
+* Updated the README example code
+
 2.10.0 - (May 23, 2018)
 ------------------
 ### Changed